requirements.txt
pandas
pyotp
python-dotenv
SmartApi
